Cleanup exception handling.
Introduces a number of new exceptions and cleans up/corrects some
probably incorrect behavior.

There is now a RedlockError base class derived from Exception from which
all other redlock specific exceptions are derived.

MultipleRedlockException now derives from RedlockError. It generally
contains a list of redis.exceptions.RedisError objects, but at Redlock
construction time, it may also contain any exception raised by
redis.StrictRedis initialization.

QuorumError deerives from MultipleRedlockException and is raised by
Redlock initialization if connections to a majority of the servers
cannot be created.

Two completely new exceptions are added: OwnerError, and MissingError.
OwnerError is raised if you attempt to operate on (unlock or extend) a
lock that you do not own (as determined by the key value).

MissingError is raised if you attempt to unlock or extend a key that
does not exist at all which would happen if you had previously unlocked
it or if the lock timed out and was removed from redis.

Given a quorum of servers, some of them may be in slightly different
states. For example, if locks are acquired on 2 redis servers
sequentially, the will expire at slightly different times. It's possible
that you might try to unlock the lock when the key on one server has
expired and the other has not.

When attempting to unlock the lock, if any of the servers is unlocked
then the unlock operation succeeds. If none of the servers succeed then
OwnerError or MissingError will be raised based on whether any of the
servers is locked with a different key or not.

An extend operation will only succeed if a quorum of the servers can be
successfully extended. Again, some of them could have expired and
possibly been (temporarily) acquired by another owner. If less than a
quorum can be extended then we've effectively lost the lock and the lock
is released immediately on all servers.

Caveat: We don't yet have tests cases for all of these corner cases.