Replace test with query, add tests, remove drift
The previous "test" method simply reported whether or not the lock was
held, but there was no useful info about who was holding it or for how
long. The new query method returns a redlock.Lock object that represents
the current state of the lock. If the lock is not currently held, query
returns Lock(0, resource, None).

Added tests for extend and query methods.

Removed the clock_drift_factor. 1% clock drift is *huge* in modern
systems and it caused the lock method to lie about the validity (ttl) of
the lock that was returned. For larger ttl values (e.g., 300000 ms) the
drift factor caused it to return a result that was several full seconds
off the actual time. The query method returns the redis server's idea of
ttl, so lock does too.

Finally, be careful to ensure that all values added to Lock objects are
binary values and not text values. For python2 there is little
difference since str is the binary type. But for python3, str is the
text type and bytes is the binary type. We do this to ensure that the
Lock that's returned from the lock() method and the Lock that's returned
from the query() method have matching values.