Teach lock to work with caller supplied keys
Add an optional parameter to lock to allow the key to be provided by the
caller. The key is the unique string that is stored in the lock value.
Knowledge of this key value allows you to unlock the resource.

This facilitates the use of redlock by cooperating processes that may
already have a shared unique piece of information but may not be able to
easily pass the key generated by redlock between each other. For
example, to acquire the lock in a pre-checkin trigger and to release it
in a post-checkin trigger.