Fix the elapsed time in lock and extend
The previous calculation stored the start and end times as milliseconds,
then calculated the elapsed time as their difference. But the start and
end time could end up with different millisecond value, even if the true
elapsed time was only a microsecond.

Now we calculate the "true" (as true as time.time() can return) value
for the elapsed time and convert that value to milliseconds. It is much
more likely to be 0 and the test that locks and then immediately queries
the key's value will now find monotonically decreasing values.